+++ KWDP/KWDP/Objects/EcgValues.cs
@@ -7,7 +7,7 @@ namespace KWDP.View
     {
         public int SampleNumber { get; set; }
         
-        public short Signal { get; set; }
+        public float Signal { get; set; }
 
         public static EcgValues FromCsv(string csvLine)
         {
@@ -15,7 +15,7 @@ namespace KWDP.View
 
             EcgValues ecg = new EcgValues();
             ecg.SampleNumber = int.Parse(values[0]);
-            ecg.Signal = ((short)float.Parse(values[1], CultureInfo.InvariantCulture));
+            ecg.Signal = (float.Parse(values[1], CultureInfo.InvariantCulture));
             return ecg;
         }
     }
